+++ components/PlayStory/AnswerBox.js
@@ -1,10 +1,12 @@
-import { View, FlatList, StyleSheet } from "react-native";
+import { View, FlatList, Text, StyleSheet } from "react-native";
 import AnswerItem from "./AnswerItem";
 import { GlobalStyles } from "../../constants/styles";
 import NextButton from "./NextButton";
 
 function AnswerBox({ answers, correct_answer_idx }) {
   const validSentence = correct_answer_idx !== -1;
+  // if answers is undefined, set to array of 4 empty strings
+  const validAnswers = answers !== null ? answers : ["", "", "", ""];
   // console.log("AnswerBox");
   // console.log(correct_answer_idx);
   // console.log(validSentence);
@@ -20,14 +22,14 @@ function AnswerBox({ answers, correct_answer_idx }) {
 
   return (
     <View style={styles.AnswerBoxContainer}>
-      {validSentence && (
+      {
         <FlatList
-          data={answers}
+          data={validAnswers}
           renderItem={renderAnswerItem}
-          keyExtractor={(index) => index}
+          keyExtractor={(item, index) => index}
           numColumns={2}
         />
-      )}
+      }
       <NextButton skip={!validSentence}></NextButton>
     </View>
   );