+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2023-XX-XX:
 - Add show totals and avoid editing > total
-- 
+- Move languages to the settings screen and filter by language
+
 
 2023-10-11:
 - Add ability to import data to filesystem (only Android)